third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-history-replaceState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  let start_length = history.length;
  appHistory.onnavigate = t.step_func(e => {
    assert_equals(e.canRespond, true);
    assert_equals(e.hashChange, false);
    assert_equals(e.formData, null);
    e.preventDefault();
    t.step_timeout(t.step_func_done(() => {
      assert_equals(location.hash, "");
      assert_equals(history.state, null);
      assert_equals(history.length, start_length);
    }), 0);
  });
  history.replaceState(1, null, "#");
}, "history.replaceState() fires the navigate event");
</script>
